Move logic into the GridIntensity object we use as a prototype
