+++ src/node.js
@@ -1,33 +1,18 @@
+// 3rd party / native libs
 import fetch from "cross-fetch"
 import { LocalStorage } from "node-localstorage"
-import settings from "./defaultSettings"
 
+// local to this project
+import settings from "./defaultSettings"
 import GridIntensityMixin from "./gridIntensity"
 
+// set up our objects
 const localStorage = new LocalStorage("./scratch")
-
 const GridIntensity = Object.create(GridIntensityMixin)
 
 GridIntensity.fetch = fetch
 GridIntensity.localStorage = localStorage
 GridIntensity.data = []
 GridIntensity.intensityProvider = settings.uk
-
-GridIntensity.fetchIntensityData = async function () {
-  const now = new Date()
-  const [before, after] = this.intensityProvider.api.forwardLooking.split(
-    "{from}"
-  )
-  const urlString = `${before}${now.toISOString()}${after}/`
-  let res = await this.fetch(urlString)
-  this.data = await res.json()
-  return this.data
-}
-
-GridIntensity.getCarbonIndex = async function (options) {
-  let now = this.getInterval(options)
-  const latestReading = await this.getReadingForInterval(now)
-  return latestReading.intensity.index
-}
 
 export default GridIntensity
